Add WIP `/microblot new` implementation
This doesn't really work yet.

The modal fires nicely and looks lovely, but submitting it does…well,
nothing, really. The spinny circle keeps spinning and spinning and
spinning and eventually "We had some trouble connecting." shows up.
In the meantime, there are no requests to the app, so either something
is misconfigured, or I'm misunderstanding how this is supposed to work.

To be fixed. >_>